change localip is6 flag based on destination ip
By default, the localip flag is initialized with
the is6 flag set to 0. The only time the flag is
changed to 1 is when a user specifies a local IPv6
address to use.

Because of this, socket_tcp46 always creates an IPv4
socket.

This patch corrects that - if the user hasn't specified
a local address, then the local 'is6' flag is updated
to match the destination 'is6' flag.

Signed-off-by: Laurent Bercot <ska-skaware@skarnet.org>